terraform/routeros/config/tf: Remove Bogon filter
Under normal circumstances you'd want to filter the Bogons at the wan
interface.  When stacking NATs, however, the next hop up is going to
be a bogon and will result in the router just dropping traffic at its
local firewall.  This has the practical effect that pure-IP links work
(since the forward chain accepts the traffic) but things on the router
itself such as DNS fail.