pkg/http: Remove location from HUD.
This commit temporarily removes location verification from the HUD.
This feature depended on checking the name provided by the device at
the other end, and everywhere else in the system we do not trust the
device at the other end.  This feature will need to be re-implemented
as a server side check to see if the expected name is being refused
publish because it is pushing from the wrong location.